safari-600.7-branch/LayoutTests/platform/ios-sim/storage/indexeddb/pending-version-change-stuck-works-with-terminate-expected.txt

CONSOLE MESSAGE: line 24: TypeError: undefined is not an object (evaluating 'request.onblocked = unexpectedBlockedCallback')
Explicitly terminating worker with blocked call should allow later open calls to proceed

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


indexedDB = self.indexedDB || self.webkitIndexedDB || self.mozIndexedDB || self.msIndexedDB || self.OIndexedDB;

dbname = "pending-version-change-stuck-works-with-terminate.html"
indexedDB.deleteDatabase(dbname)
FAIL indexedDB.deleteDatabase(dbname) threw exception TypeError: null is not an object (evaluating 'indexedDB.deleteDatabase')
FAIL Unexpected error: TypeError: undefined is not an object (evaluating 'request.onblocked = unexpectedBlockedCallback')
FAIL successfullyParsed should be true. Was false.

TEST COMPLETE
